+++ workflows/codefresh-csdp/versions/0.0.6/images/report-image-info/index.js
@@ -2,42 +2,147 @@ const { GraphQLClient, gql } = require('graphql-request')
 const fs = require('fs')
 const _ = require('lodash')
 const AWS = require('aws-sdk');
+const { parseQualifiedNameOptimized, parseFamiliarName } = require('@codefresh-io/docker-reference')
 
 const { registries: { GcrRegistry, EcrRegistry, DockerhubRegistry, StandardRegistry } } = require('nodegistry');
 
+// Trim all input
+// Clean this up to use the same variable with a 'registry type'
+const inputs = require('./configuration');
+
 const CF_NOT_EXIST = 'cf-not-exist';
 
-async function createRegistryClient() {
+function checkNotEmpty(testVar) {
+    return (testVar && testVar!==CF_NOT_EXIST);
+}
 
-    if (process.env.DOCKER_USERNAME && process.env.DOCKER_PASSWORD
-        && process.env.DOCKER_USERNAME!==CF_NOT_EXIST && process.env.DOCKER_PASSWORD!==CF_NOT_EXIST) {
+function _parseImageName(imageName) {
+    return parseFamiliarName(imageName, parseQualifiedNameOptimized)
+}
+
+const _decodeBase64 = (str) => Buffer.from(str, 'base64').toString();
+
+function getCredentialsFromDockerConfig(image) {
+    const dockerConfig = JSON.parse(fs.readFileSync(inputs.dockerConfigPath));
+    const imageData = _parseImageName(image);
+    const auths = _.get(dockerConfig, 'auths', {});
+    const domainKey = _.findKey(auths, (auth, domain) => {
+        if (domain.includes(imageData.domain)) {
+            return true
+        }
+    });
+    const authInfo = auths[domainKey];
+    const auth = _decodeBase64(authInfo.auth);
+    const [ username, password ] = auth.split(':');
+    if (domainKey.includes('docker.io')) {
         return new DockerhubRegistry({
-            username: process.env.DOCKER_USERNAME,
-            password: process.env.DOCKER_PASSWORD
+            username,
+            password,
         });
+    } else if (domainKey.includes('azurecr.io')) {
+        throw new Error('Azure Container Registry using docker config json is not supported.')
+    } else if (domainKey.includes('gcr.io')) {
+        throw new Error('Google container registry using docker config json is not supported.')
+    }
+    return new StandardRegistry({
+        request: {
+            host: domainKey
+        },
+        credentials: {
+            username,
+            password,
+        },
+    });
+}
+
+async function createRegistryClientByImage(image) {
+    if (checkNotEmpty(inputs.dockerConfigPath)) {
+        return getCredentialsFromDockerConfig(image)
+        return getCredentialsFromDockerConfig(image)
+    }
+    const imageData = _parseImageName(image);
+    if (imageData.domain.includes('docker.io')) {
+        if (checkNotEmpty(inputs.docker.username)
+            && checkNotEmpty(inputs.docker.password)) {
+
+            return new DockerhubRegistry(inputs.docker);
+        }
+        throw new Error('Registry credentials for DOCKER not set. Add following registry parameters in your workflow to continue:\n - DOCKER_USERNAME\n - DOCKER_PASSWORD\n');
+    } else if (imageData.domain.includes('gcr.io')) {
+        if (inputs.gcr.keyFilePath) {
+            return new GcrRegistry({
+                keyfile: fs.readFileSync(inputs.gcr.keyFilePath),
+                request: { host: 'gcr.io' }
+            });
+        }
+        throw new Error('Registry credentials for GCR not set. Add following registry parameters in your workflow to continue:\n - GCR_KEY_FILE_PATH\n');
+    } else if (imageData.domain.includes('azurecr.io')) {
+        throw new Error('Azure Container Registry using docker config json is not supported.')
+    } else if (imageData.domain.includes('ecr')) {
+        if (checkNotEmpty(inputs.aws.role)
+            && checkNotEmpty(inputs.aws.credentials.region)) {
+            console.log(`Retrieving credentials for ECR ${inputs.aws.region} using STS token`);
+            const sts = new AWS.STS();
+            const timestamp = (new Date()).getTime();
+            const params = {
+                RoleArn: inputs.aws.role,
+                RoleSessionName: `be-descriptibe-here-${timestamp}`
+            }
+            const data = await sts.assumeRole(params).promise();
+            return new EcrRegistry({
+                promise: Promise,
+                credentials: {
+                    accessKeyId: data.Credentials.AccessKeyId,
+                    secretAccessKey: data.Credentials.SecretAccessKey,
+                    sessionToken: data.Credentials.SessionToken,
+                    region: inputs.aws.credentials.region,
+                },
+            })
+        } else if (checkNotEmpty(inputs.aws.credentials.accessKeyId)
+            && checkNotEmpty(inputs.aws.credentials.secretAccessKey)
+            && checkNotEmpty(inputs.aws.credentials.region)) {
+            return new EcrRegistry({
+                promise: Promise,
+                credentials: inputs.aws.credentials,
+            })
+        }
+        throw new Error('Registry credentials for ECR not set. Add following registry parameters in your workflow to continue:\n - AWS_ACCESS_KEY\n - AWS_SECRET_KEY\n - AWS_REGION\n');
+    }
+    if (checkNotEmpty(inputs.generic.credentials.username)
+        && checkNotEmpty(inputs.generic.credentials.password)
+        && checkNotEmpty(inputs.generic.request.host)) {
+        return new StandardRegistry(inputs.generic);
+    }
+    throw new Error('Registry credentials is required parameter. Add one from following registry parameters in your workflow to continue:\n - Docker credentials: DOCKER_USERNAME, DOCKER_PASSWORD\n - GCR credentials: GCR_KEY_FILE_PATH\n - AWS registry credentials: AWS_ACCESS_KEY, AWS_SECRET_KEY, AWS_REGION\n - Standard registry credentials: USERNAME, PASSWORD, DOMAIN\n')
+}
+
+async function createRegistryClient(image) {
+
+    if (checkNotEmpty(inputs.dockerConfigPath)) {
+        return getCredentialsFromDockerConfig(image)
     }
 
-    if (process.env.USERNAME && process.env.PASSWORD && process.env.DOMAIN
-        && process.env.USERNAME!==CF_NOT_EXIST && process.env.PASSWORD!==CF_NOT_EXIST && process.env.DOMAIN!==CF_NOT_EXIST) {
-        return new StandardRegistry({
-            request: {
-                protocol: process.env.INSECURE === 'true' ? 'http' : 'https',
-                host: process.env.DOMAIN
-            },
-            credentials: {
-                username: process.env.USERNAME,
-                password: process.env.PASSWORD,
-            },
-        });
+    // Clean this up when have time
+    if (checkNotEmpty(inputs.docker.username)
+        && checkNotEmpty(inputs.docker.password)) {
+
+        return new DockerhubRegistry(inputs.docker);
     }
 
-    if (process.env.AWS_ROLE && process.env.AWS_ROLE!==CF_NOT_EXIST
-        && process.env.AWS_REGION && process.env.AWS_REGION!==CF_NOT_EXIST) {
-        console.log(`Retrieving credentials for ECR ${process.env.AWS_REGION} using STS token`);
+    if (checkNotEmpty(inputs.generic.credentials.username)
+        && checkNotEmpty(inputs.generic.credentials.password)
+        && checkNotEmpty(inputs.generic.request.host)) {
+
+            return new StandardRegistry(inputs.generic);
+    }
+
+    if (checkNotEmpty(inputs.aws.role)
+        && checkNotEmpty(inputs.aws.credentials.region)) {
+        console.log(`Retrieving credentials for ECR ${inputs.aws.region} using STS token`);
         const sts = new AWS.STS();
         const timestamp = (new Date()).getTime();
         const params = {
-            RoleArn: process.env.AWS_ROLE,
+            RoleArn: inputs.aws.role,
             RoleSessionName: `be-descriptibe-here-${timestamp}`
         }
         const data = await sts.assumeRole(params).promise();
@@ -47,39 +152,42 @@ async function createRegistryClient() {
                 accessKeyId: data.Credentials.AccessKeyId,
                 secretAccessKey: data.Credentials.SecretAccessKey,
                 sessionToken: data.Credentials.SessionToken,
-                region: process.env.AWS_REGION,
+                region: inputs.aws.credentials.region,
             },
         })
     }
 
-    if (process.env.GCR_KEY_FILE_PATH) {
+    if (inputs.gcr.keyFilePath) {
         return new GcrRegistry({
-            keyfile: fs.readFileSync(process.env.GCR_KEY_FILE_PATH),
+            keyfile: fs.readFileSync(inputs.gcr.keyFilePath),
             request: { host: 'gcr.io' }
         });
     }
-    if (process.env.AWS_ACCESS_KEY && process.env.AWS_ACCESS_KEY!==CF_NOT_EXIST
-        && process.env.AWS_SECRET_KEY && process.env.AWS_SECRET_KEY!==CF_NOT_EXIST
-        && process.env.AWS_REGION && process.env.AWS_REGION!==CF_NOT_EXIST) {
+    if (checkNotEmpty(inputs.aws.credentials.accessKeyId)
+        && checkNotEmpty(inputs.aws.credentials.secretAccessKey)
+        && checkNotEmpty(inputs.aws.credentials.region)) {
         return new EcrRegistry({
             promise: Promise,
-            credentials: {
-                accessKeyId: process.env.AWS_ACCESS_KEY,
-                secretAccessKey: process.env.AWS_SECRET_KEY,
-                region: process.env.AWS_REGION,
-            },
+            credentials: inputs.aws.credentials,
         })
     }
     throw new Error('Registry credentials is required parameter. Add one from following registry parameters in your workflow to continue:\n - Docker credentials: DOCKER_USERNAME, DOCKER_PASSWORD\n - GCR credentials: GCR_KEY_FILE_PATH\n - AWS registry credentials: AWS_ACCESS_KEY, AWS_SECRET_KEY, AWS_REGION\n - Standard registry credentials: USERNAME, PASSWORD, DOMAIN');
 }
 
+async function getRegistryClient(image) {
+    if (inputs.retrieveCredentialsByDomain) {
+        return createRegistryClientByImage(image);
+    }
+    return createRegistryClient(image);
+}
+
 const init = async () => {
 
-    const client = await createRegistryClient();
+    const image = inputs.image.uri;
+    const client = await getRegistryClient(image);
 
-    const image = process.env.IMAGE_URI;
-    const authorUserName = process.env.GIT_SENDER_LOGIN;
-    const workflowName = process.env.WORKFLOW_NAME;
+    const authorUserName = inputs.git.author;
+    const workflowName = inputs.workflow.name;
 
     const registry = client.repoTag(image);
 
@@ -90,9 +198,9 @@ const init = async () => {
         return sum + layer.size;
     }, 0)
 
-    const graphQLClient = new GraphQLClient(`${process.env.CF_HOST}/2.0/api/graphql`, {
+    const graphQLClient = new GraphQLClient(`${inputs.codefresh.host}/2.0/api/graphql`, {
         headers: {
-            'Authorization': process.env.CF_API_KEY,
+            'Authorization': inputs.codefresh.apiKey,
         },
     })
 
@@ -100,10 +208,10 @@ const init = async () => {
         id: manifest.config.digest,
         created: config.created,
         imageName: image,
-        branch: process.env.GIT_BRANCH,
-        commit: process.env.GIT_REVISION,
-        commitMsg: process.env.GIT_COMMIT_MESSAGE,
-        commitURL: process.env.GIT_COMMIT_URL,
+        branch: inputs.git.branch,
+        commit: inputs.git.commit,
+        commitMsg: inputs.git.commitMsg,
+        commitURL: inputs.git.commitURL,
         size: size,
         os: config.os,
         architecture: config.architecture,
@@ -149,10 +257,10 @@ const init = async () => {
 }
 
 const validateRequiredEnvs = () => {
-    if (_.isEmpty(process.env.IMAGE_URI)) {
+    if (_.isEmpty(inputs.image.uri)) {
         throw new Error('IMAGE_URI is required parameter. Add this parameter in your workflow to continue.');
     }
-    if (_.isEmpty(process.env.CF_API_KEY)) {
+    if (_.isEmpty(inputs.codefresh.apiKey)) {
         throw new Error('CF_API_KEY is required parameter. Add this parameter in your workflow to continue.');
     }
 }